backend_backup/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
google-generativeai
pytest
httpx
